OMAP4470: Fatal threshold hardware warm reset generation
On 4470 the TSHUT signal can be mapped to the device SCRM external
warm reset logic to activate the PRM.SYS_WARMIN_RST input
when a fatal temperature threshold (TSHUT_HOT) is exceeded.
Hence, a hardware warm reset can be triggered directly on
the device upon a fatal temperature threshold crossing event.

This patch adds support of the above described feature.

Change-Id: I236bdfd1669838bbf085cc5f7ce0e60161f0c8ec
Signed-off-by: Leonid Iziumtsev <x0153368@ti.com>